prevent auto-scroll to bottom if we're not at the bottom
Also prevent sending of blank messages by holding Enter.

In truth, this should probably be a user setting somewhere. Some people
want to scroll to the bottom every time. I think.

Also, I'd like this to be bottom-5 lines or something like that, but I
only get percentage amount for the scroll bar. So, scrolling so the
bottom empty line is hidden will probably disable the autoscroll.