Add json and msgpack support, and commands.
Modes can be switched with commands. The command tells your peer that
you will be sending new messages in the new format. Commands don't
affect what format your peer sends messages in. Also, /json or /msgpack
command must be used for the header (to/from) to be sent.

Header "from" is used to show a nicer peer address, like a domain for
future contact.

Current commands:
/text - text mode (default)
/json - json mode
/msgpack - msgpack mode
/bye - instruct your peer to close the TCP connection

/bye was mostly for testing. It may be considered an application-level
disconnection rather than closing the window, which is a TCP-level
disconnect (FIN).

Probably future versions should default to msgpack, and we may not want
to keep json anyway.